changed variables, fixed small bugs

+++ src/org/hikst/Simulator/Object.java
@@ -16,13 +16,13 @@ public class Object
 	private float current;
 	private int impact_degree_ID;
 	private int usage_pattern_ID;
-	private double latitude;
-	private double longitude;
-	private double self_temperature;
-	private double target_temperature;
-	private double base_area;
-	private double base_height;
-	private double heat_loss_rate;
+	private float latitude;
+	private float longitude;
+	private float self_temperature;
+	private float target_temperature;
+	private float base_area;
+	private float base_height;
+	private float heat_loss_rate;
 	private ArrayList<Integer> sons = new ArrayList<Integer>();
 	
 	public int getID() {
@@ -53,31 +53,31 @@ public class Object
 		return usage_pattern_ID;
 	}
 	
-	public double getLatitude(){
+	public float getLatitude(){
 		return latitude;
 	}
 	
-	public double getLongitude(){
+	public float getLongitude(){
 		return longitude;
 	}
 	
-	public double getSelfTemperature(){
+	public float getSelfTemperature(){
 		return self_temperature;
 	}
 	
-	public double getTargetTemperature(){
+	public float getTargetTemperature(){
 		return target_temperature;
 	}
 	
-	public double getBaseArea(){
+	public float getBaseArea(){
 		return base_area;
 	}
 	
-	public double getBaseHeight(){
+	public float getBaseHeight(){
 		return base_height;
 	}
 	
-	public double getHeatLossRate(){
+	public float getHeatLossRate(){
 		return heat_loss_rate;
 	}
 	
@@ -115,12 +115,12 @@ public class Object
 				this.current = set.getFloat(5);
 				this.impact_degree_ID = set.getInt(6);
 				this.usage_pattern_ID = set.getInt(7);
-				this.longitude = set.getDouble(8);
-				this.latitude = set.getDouble(9);
-				this.self_temperature = set.getDouble(10);
-				this.target_temperature = set.getDouble(11);
-				this.base_area = set.getDouble(12);
-				this.base_height = set.getDouble(13);
+				this.longitude = set.getFloat(8);
+				this.latitude = set.getFloat(9);
+				this.self_temperature = set.getFloat(10);
+				this.target_temperature = set.getFloat(11);
+				this.base_area = set.getFloat(12);
+				this.base_height = set.getFloat(13);
 				
 				query = "SELECT Son_ID FROM Part_Objects WHERE Father_ID=?";
 				PreparedStatement anotherStatement = connection.prepareStatement(query);
